Why does it not work? What is the reason? The weather is okish today. Should I go to swim or not? I have to make up to long fake commit message..... This week we only work for thress days. My plan is to finish my current story and if it is possible try to start a new article about music. I do not know if it is right to write about these. It is like telling my own stories to the public. I do not feel so comfortable. Maybe that is the reason that i could not fall asleep
Why does it not work? What is the reason? The weather is okish today. Should I go to swim or not? I have to make up to long fake commit message..... This week we only work for thress days. My plan is to finish my current story and if it is possible try to start a new article about music. I do not know if it is right to write about these. It is like telling my own stories to the public. I do not feel so comfortable. Maybe that is the reason that i could not fall asleep

Why does it not work? What is the reason? The weather is okish today. Should I go to swim or not? I have to make up to long fake commit message..... This week we only work for thress days. My plan is to finish my current story and if it is possible try to start a new article about music. I do not know if it is right to write about these. It is like telling my own stories to the public. I do not feel so comfortable. Maybe that is the reason that i could not fall asleep